refactor(ui): Application, Controller 분리
